requirements.txt
anyio==3.6.1
asgiref==3.5.2
azure-common==1.1.28
azure-core==1.24.1
azure-cosmos==4.3.0
azure-cosmosdb-nspkg==2.0.2
azure-cosmosdb-table==1.0.6
azure-identity==1.10.0
azure-keyvault-secrets==4.4.0
azure-nspkg==3.0.2
brotlipy==0.7.0
certifi==2021.10.8
cffi @ file:///C:/ci_310/cffi_1642682485096/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==7.1.2
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
conda==4.12.0
conda-content-trust @ file:///tmp/build/80754af9/conda-content-trust_1617045594566/work
conda-package-handling @ file:///C:/ci/conda-package-handling_1649106011304/work
cryptography @ file:///C:/ci/cryptography_1639414720302/work
distlib==0.3.4
distro==1.7.0
dnspython==2.2.1
email-validator==1.1.2
fastapi==0.78.0
filelock==3.7.1
gunicorn==20.1.0
h11==0.13.0
httptools==0.4.0
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
influxdb-client==1.24.0
isodate==0.6.1
menuinst @ file:///C:/ci/menuinst_1631733438520/work
motor==2.0.0
msal==1.18.0
msal-extensions==1.0.0
msrest==0.7.1
ntlm-auth==1.5.0
numpy==1.22.4
oauthlib==3.2.0
osisoft.pidevclub.piwebapi @ https://github.com/dcbark01/PI-Web-API-Client-Python/archive/master.zip
pandas==1.4.2
pipenv==2022.6.7
platformdirs==2.5.2
portalocker==2.4.0
psutil==5.9.1
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==1.9.1
PyJWT==1.7.1
pymongo==3.10.1
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
PySocks @ file:///C:/ci/pysocks_1605307512533/work
pyspnego==0.5.2
python-dateutil==2.8.2
python-decouple==3.3
python-dotenv==0.20.0
python-multipart==0.0.5
pytz==2022.1
pywin32==302
PyYAML==6.0
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
requests-kerberos==0.14.0
requests-ntlm==1.1.0
requests-oauthlib==1.3.1
ruamel-yaml-conda @ file:///C:/ci/ruamel_yaml_1616016898638/work
Rx==3.2.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.2.0
starlette==0.19.1
tabula-py==2.4.0
tqdm @ file:///opt/conda/conda-bld/tqdm_1647339053476/work
typing_extensions==4.2.0
unicorn==1.0.3
urllib3 @ file:///opt/conda/conda-bld/urllib3_1643638302206/work
uvicorn==0.18.1
virtualenv==20.14.1
virtualenv-clone==0.5.7
watchgod==0.8.2
websockets==10.3
Werkzeug==2.1.2
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306162074/work
wincertstore==0.2
wrapt==1.14.1
